creates a new map based off of the old map implementation to avoid unnecesary functions/styling/misunderstandings with naming conventions
