+++ package.json
@@ -4,6 +4,7 @@
   "private": false,
   "dependencies": {
     "@craco/craco": "^5.6.4",
+    "@mapbox/geo-viewport": "^0.4.1",
     "@material-ui/core": "^4.11.0",
     "@material-ui/icons": "^4.9.1",
     "@material-ui/styles": "^4.10.0",
